chore: create page layouts and auth
