+++ layouts/auth.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts"></script>
+
+<template>
+  <div class="flex flex-col justify-center items-center gap-5 text-center h-screen px-4">
+    <slot />
+  </div>
+</template>
+
+<style scoped></style>
