+++ middleware/auth.ts
@@ -1,0 +1,9 @@
+export default defineNuxtRouteMiddleware(() => {
+  const isAuthenticated = false;
+  // If the user is not authenticated, redirect to the auth page
+  if (!isAuthenticated) {
+    return navigateTo({
+      name: "auth"
+    });
+  }
+});
